refactor: positionToKeyをpositionUtilsに移動し循環依存を解消
